[CodeCompletion] Move TypeCheckCompletionCallbacks to IDE
