+++ include/swift/IDE/TypeCheckCompletionCallback.h
@@ -1,0 +1,81 @@
+//===--- TypeCheckCompletionCallback.h  -----------------------------------===//
+//
+// Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+//
+/// \file
+/// Provides TypeCheckCompletionCallback implementations for the various kinds
+/// of code completion. These extract and persist information needed to compute
+/// completion results from the solutions formed during expression typechecking.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef SWIFT_IDE_TYPECHECKCOMPLETIONCALLBACK_H
+#define SWIFT_IDE_TYPECHECKCOMPLETIONCALLBACK_H
+
+#include "swift/AST/Expr.h"
+#include "swift/AST/Type.h"
+#include "swift/Basic/LLVM.h"
+#include "llvm/ADT/DenseMap.h"
+#include "llvm/ADT/SmallVector.h"
+
+namespace swift {
+class Decl;
+class DeclContext;
+class Type;
+class ValueDecl;
+class CodeCompletionExpr;
+
+namespace constraints {
+class ConstraintSystem;
+class Solution;
+} // namespace constraints
+
+namespace ide {
+
+class TypeCheckCompletionCallback {
+  bool GotCallback = false;
+
+public:
+  virtual ~TypeCheckCompletionCallback() {}
+
+  /// Called for each solution produced while type-checking an expression
+  /// that the code completion expression participates in.
+  virtual void sawSolution(const constraints::Solution &solution) {
+    GotCallback = true;
+  };
+
+  /// True if at least one solution was passed via the \c sawSolution
+  /// callback.
+  bool gotCallback() const { return GotCallback; }
+
+  /// Typecheck the code completion expression in its outermost expression
+  /// context, calling \c sawSolution for each solution formed.
+  virtual void fallbackTypeCheck(DeclContext *DC);
+};
+
+// MARK: - Utility functions for subclasses of TypeCheckCompletionCallback
+
+Type getTypeForCompletion(const constraints::Solution &S, Expr *E);
+
+/// Whether the given completion expression is the only expression in its
+/// containing closure or function body and its value is implicitly returned.
+///
+/// If these conditions are met, code completion needs to avoid penalizing
+/// completion results that don't match the expected return type when
+/// computing type relations, as since no return statement was explicitly
+/// written by the user, it's possible they intend the single expression not
+/// as the return value but merely the first entry in a multi-statement body
+/// they just haven't finished writing yet.
+bool isImplicitSingleExpressionReturn(constraints::ConstraintSystem &CS,
+                                      Expr *CompletionExpr);
+
+} // namespace ide
+} // namespace swift
+
+#endif // SWIFT_IDE_TYPECHECKCOMPLETIONCALLBACK_H
